Add work in progress on African Easterly Waves component class
AEW class uses windspharm module to calculate vorticity so
environment.yml is updated to specify this module.

A global domain is required for the vorticity calculations, so an
argument has been added to SynopticChart::get_data() to indicate
whether the chart's domain constraint should be applied.

Initially, we're locating AEW troughs/ridges where advection of the
stream-function vorticity by the nondivergent wind is equal to zero.
This needs to be updated to use advection of the curvature component
of the vorticity.